// src/interfaces/IState.ts
export interface IState {
    articles: string;
    articles1: string;
    articles2: string;
    articles3: string;
    unidades: string;
    unidades1: string;
    unidades2: string;
    unidades3: string;
    precio: number;
    precio1: number;
    precio2: number;
    precio3: number;
    cesta: any;
    cesta1: any;
    cesta2: any;
    cesta3: any;
    impuesto: number;
    total: number;
}